requirement.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
exceptiongroup==1.2.2
fastmcp==2.2.7
filelock==3.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
identify==2.6.10
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
markdown-it-py==3.0.0
mcp==1.7.1
mdurl==0.1.2
nodeenv==1.9.1
openapi-pydantic==0.5.1
packaging==25.0
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
virtualenv==20.30.0
websockets==15.0.1
google-generativeai
mcp-client # Added for testing the server
